pipeline/requirements.txt
pymupdf>=1.24.0
python-pptx>=1.0.0
pillow>=10.0.0
pydantic>=2.7.0
groq>=0.7.0
httpx>=0.27.0
python-dotenv>=1.0.0
tqdm>=4.66.0
